Add files via upload
This file contains row echelon form of a matrix. Input format: Input the dimension of the original nxn matrix. Then input the number of columns of the augmented matrix (say m) (you may enter 0 if no augmentation is required) in a new line. Then enter the values of the matrix row-wise including the rows of the augmentation matrix. Then the output would be step by step row-echelon form solver and finally the reduced-row echelon form of the matrix.